+++ src/include/coati/mutation_coati.hpp
@@ -32,11 +32,11 @@
 
 #include "utils.hpp"
 
-using Vector64f = Eigen::Matrix<double, 64, 1>;
-using Matrix4f = Eigen::Matrix<double, 4, 4>;
+using Vector64f = Eigen::Matrix<float, 64, 1>;
+using Matrix4f = Eigen::Matrix<float, 4, 4>;
 
 void mg94_q(Matrix64f& Q);
-void mg94_p(Matrix64f& P, double br_len);
-void mg94_marginal_p(Eigen::Tensor<double, 3>& p, Matrix64f& P);
+void mg94_p(Matrix64f& P, float br_len);
+void mg94_marginal_p(Eigen::Tensor<float, 3>& p, Matrix64f& P);
 
 #endif
